space from waterhole good
I need to make sure the water holes are not hidden behind rocks so that the player can get them. Now there is a possibility for a buffer between all rocks and a water hole and it is up to game design to determine how large this space should be. This functionality should not make the game slow at all